@@ -17,7 +17,7 @@
         }
     ],
     "require": {
-        "php": ">=7.4",
+        "php": ">=8.0",
         "symfony/config": "^5.4 || ^6.0",
         "symfony/dependency-injection": "^5.4 || ^6.0",
         "symfony/form": "^5.4 || ^6.0",
@@ -35,6 +35,7 @@
         "phpspec/phpspec": "^7.2",
         "phpspec/prophecy": "^1.15",
         "phpstan/phpstan": "^1.9",
+        "rector/rector": "^0.14.7",
         "symfony/twig-bridge": "^5.4 || ^6.0",
         "symfony/var-dumper": "^5.4 || ^6.0"
     },
