@@ -16,11 +16,8 @@ final class Combined implements Factory
      */
     private const TYPE = 'combined';
 
-    private Dictionary\Collection $dictionaries;
-
-    public function __construct(Collection $dictionaries)
+    public function __construct(private Collection $dictionaries)
     {
-        $this->dictionaries = $dictionaries;
     }
 
     public function create(string $name, array $config): Dictionary
