@@ -11,11 +11,8 @@ use Psr\Container\ContainerInterface;
 
 final class Invokable implements Factory
 {
-    private ContainerInterface $container;
-
-    public function __construct(ContainerInterface $container)
+    public function __construct(private ContainerInterface $container)
     {
-        $this->container = $container;
     }
 
     /**
