@@ -12,11 +12,8 @@ use Traversable;
 
 final class Iterator implements Factory
 {
-    private ContainerInterface $container;
-
-    public function __construct(ContainerInterface $container)
+    public function __construct(private ContainerInterface $container)
     {
-        $this->container = $container;
     }
 
     /**
