@@ -15,16 +15,10 @@ use ReturnTypeWillChange;
 abstract class Wrapper implements Dictionary
 {
     /**
-     * @var Dictionary<E>
-     */
-    private Dictionary $wrapped;
-
-    /**
      * @param Dictionary<E> $wrapped
      */
-    public function __construct(Dictionary $wrapped)
+    public function __construct(private Dictionary $wrapped)
     {
-        $this->wrapped = $wrapped;
     }
 
     public function getName(): string
