@@ -12,17 +12,11 @@ final class DictionaryValidator extends ConstraintValidator
 {
     use DictionaryValidator\SymfonyCompatibilityTrait;
 
-    private Collection $dictionaries;
-
-    public function __construct(Collection $dictionaries)
+    public function __construct(private Collection $dictionaries)
     {
-        $this->dictionaries = $dictionaries;
     }
 
-    /**
-     * @param mixed $var
-     */
-    private function varToString($var): string
+    private function varToString(mixed $var): string
     {
         if (null === $var) {
             return 'null';
